add comment

+++ resources/views/post/comments.blade.php
@@ -1,15 +1,19 @@
 @forelse($comments as $comment)
     <div class="" style="margin: 5px 0 25px 0">
         <div class="pull-left">
-            <a href="{{ route('user.show',$comment->user->name) }}">
-                <img width="48px" height="48px" class="img-circle" src="https://static.lufficc.com/image/e0b3974866e759a93e5d41a669dc2b61.png">
+            <?php
+            $href = $comment->user ? route('user.show', $comment->username) : 'javascript:void(0);';
+            ?>
+            <a href="{{ $href }}">
+                <img width="48px" height="48px" class="img-circle"
+                     src="https://static.lufficc.com/image/e0b3974866e759a93e5d41a669dc2b61.png">
             </a>
         </div>
         <div class="comment-info" style="margin-left: 66px">
             <div style="font-size: 1.2em">
-                <a href="{{ route('user.show',$comment->user->name) }}">{{ $comment->user->name }}</a>
+                <a href="{{ $href }}">{{ $comment->username }}</a>
                 <span class="pull-right" style="color: #d0d0d0">
-                    @if(auth()->id() == $comment->user_id)
+                    @if($comment->user && auth()->id() == $comment->user_id)
                         <a data-method="delete"
                            data-modal-target="确定删除评论吗?"
                            href="javascript:void (0);"
@@ -17,7 +21,7 @@
                             <i class="fa fa-trash-o fa-fw"></i>
                         </a>
                     @endif
-                    <a href="javascript:void (0);" onclick="replySomeone('{{ $comment->user->name }}')">
+                    <a href="javascript:void (0);" onclick="replySomeone('{{ $comment->username }}')">
                         <i class="fa fa-reply fa-fw"></i>
                     </a>
             </span>
